# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareNexus API — Connected Care. Better Outcomes.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentList"
    post:
      operationId: createAssignment
      tags: [assignments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Communications ────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: string
        - name: clinicId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
    post:
      operationId: createAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/complete:
    post:
      operationId: completeAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/cancel:
    post:
      operationId: cancelAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/no-show:
    post:
      operationId: noShowAppointment
      tags: [appointments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: No Show
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ── Consultations ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    Meta:
      type: object
      required: [total, page, limit, pages]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Area:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Clinic:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
        area:
          $ref: "#/components/schemas/Area"
...
    Role:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        isSystem:
          type: boolean
        createdAt:
          type: string
...
    User:
      type: object
      required: [id, email, firstName, lastName, status, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        mobile:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
        role:
          $ref: "#/components/schemas/Role"
...
    Program:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Patient:
      type: object
      required: [id, nhsNumber, firstName, lastName, mobile, status, createdAt]
      properties:
        id:
          type: string
        nhsNumber:
          type: string
        title:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        mobile:
          type: string
        altMobile:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        ethnicity:
          type: ["string", "null"]
        patientGroup:
          type: ["string", "null"]
        userType:
          type: ["string", "null"]
        status:
          type: string
        isTest:
          type: boolean
        createdAt:
          type: string
        program:
          $ref: "#/components/schemas/Program"
        clinic:
          $ref: "#/components/schemas/Clinic"
        area:
          $ref: "#/components/schemas/Area"
...
    Assignment:
      type: object
      required: [id, patientId, doctorId, clinicId, areaId, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        clinicId:
          type: string
        areaId:
          type: string
        isTemp:
          type: boolean
        createdAt:
          type: string
        doctor:
          $ref: "#/components/schemas/User"
        patient:
          $ref: "#/components/schemas/Patient"
        clinic:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        area:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
    AssignmentInput:
      type: object
      required: [patientId, doctorId, clinicId, areaId]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        clinicId:
          type: string
        areaId:
          type: string
        isTemp:
          type: boolean
    AssignmentList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"
        meta:
          $ref: "#/components/schemas/Meta"

    # ── Communication ───────────────────────────────────────────────────────
...
    Appointment:
      type: object
      required: [id, tenantId, patientId, doctorId, clinicId, appointmentDate, durationMinutes, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        clinicId:
          type: string
        appointmentDate:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [SCHEDULED, COMPLETED, CANCELLED, NO_SHOW]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patient:
          $ref: "#/components/schemas/Patient"
        doctor:
          $ref: "#/components/schemas/User"
        clinic:
          $ref: "#/components/schemas/Clinic"
    AppointmentList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        meta:
          $ref: "#/components/schemas/Meta"
    AppointmentInput:
      type: object
      required: [patientId, doctorId, clinicId, appointmentDate]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        clinicId:
          type: string
        appointmentDate:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        appointmentDate:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        notes:
          type: string
...